acervo/cadastro: corrige comparacao enum x int, valida caracteres e fecha path traversal no upload
- AcervoController.GetNovoCodigoAceca: idMarcaAcervo.Equals(EAcervo.Geral) comparava
  int com enum e nunca era true (Object.Equals falha por tipo diferente), fazendo o
  fluxo sempre cair no else mesmo com acervo != Geral; corrigido para comparar com
  (int)EAcervo.Geral
- AuthController: mesmo bug em SocioPerfilId.Equals(EPerfil.Banido) (4 ocorrencias) -
  usuarios banidos nunca eram bloqueados no login/esqueci-senha; corrigido
- GetNovoCodigoAceca/GetFullByIdFase: IsNullOrEmpty -> IsNullOrWhiteSpace (string so
  de espacos passava e quebrava em Trim()[0])
- Validacao de caracteres do nome/codigo (letra/numero no inicio-fim, traco/espaco/
  underline no meio) replicada no servidor (GetNovoCodigoAceca e Create), alem do
  client-side ja existente em admin-acervo-cadastro.js
- Create passa a exigir CodigoAceca preenchido antes de salvar
- UploadImg: sanitiza CodigoAceca (bloqueia path traversal via "../"), valida os
  magic bytes do arquivo contra a extensao informada, para de forcar extensao
  ".jpg" sem converter o arquivo, corrige checagem de sucesso do upload local (
  testava um arquivo temporario aleatorio em vez do arquivo salvo) e passa a
  bloquear sobrescrita local (mesmo comportamento ja existente no modo FTP)
- Antiforgery token no form de Cadastro + [ValidateAntiForgeryToken] em Create
- admin-acervo-cadastro.js: valida on blur/Enter/Tab de Nome e Codigo Variante,
  remove duplicacao de show() em cmbPop_MarcaVariante e codigo morto (fn_Modal/
  fnItem_Edit_CarregarDados) que referenciava elementos inexistentes na tela
- admin-acervo-listagem.js: ao desmarcar a fase, o grid e os combos de Tipo/
  SubTipo nao eram limpos/resetados; agora reseta os combos (sem disparar o
  popup de "selecione uma fase") e limpa/esconde o grid

+++ CORE/Aceca.Adm/Controllers/AuthController.cs
@@ -245,7 +245,7 @@ namespace Aceca.Adm.Controllers
                 if (user is null)
                     return Ok(new { bResult = false, type = "ERRO", message = "Usuário Inválido" });
 
-                if (user.Socio.SocioPerfilId.Equals(EPerfil.Banido))
+                if (user.Socio.SocioPerfilId.Equals((int)EPerfil.Banido))
                     return Ok(new { bResult = false, type = "ERRO", message = "Usuário Banido" });
 
                 if (!user.Socio.Ativo)
@@ -379,7 +379,7 @@ namespace Aceca.Adm.Controllers
                 if (user is null)
                     return Ok(new { bResult = true, message = "Se o e-mail existir, você receberá as instruções." });
 
-                if (user.Socio.SocioPerfilId.Equals(EPerfil.Banido))
+                if (user.Socio.SocioPerfilId.Equals((int)EPerfil.Banido))
                     return Ok(new { bResult = false, type = "ERRO", message = "Usuário Banido" });
 
                 if (!user.Socio.Ativo)
@@ -458,7 +458,7 @@ namespace Aceca.Adm.Controllers
                 if (user is null)
                     return Ok(new { bResult = false, type = "ERRO", message = "Usuário Inválido" });
 
-                if (user.Socio.SocioPerfilId.Equals(EPerfil.Banido))
+                if (user.Socio.SocioPerfilId.Equals((int)EPerfil.Banido))
                     return Ok(new { bResult = false, type = "ERRO", message = "Usuário Banido" });
 
                 if (!user.Socio.Ativo)
@@ -557,7 +557,7 @@ namespace Aceca.Adm.Controllers
                 if (user is null)
                     return Ok(new { bResult = false, type = "ERRO", message = "Usuário Inválido" });
 
-                if (user.Socio.SocioPerfilId.Equals(EPerfil.Banido))
+                if (user.Socio.SocioPerfilId.Equals((int)EPerfil.Banido))
                     return Ok(new { bResult = false, type = "ERRO", message = "Usuário Banido" });
 
                 if (!user.Socio.Ativo)
